New bootstrap is added in the Index.cshtml file and the BuyBook.cshtml file
BOOTSTRAP ADDED #1: Table in the "cart" section of the website has a "table-hover" feature added so that when a user has their mouse over a certain row on the table, it darkens so they know that is the row they are on. Along with this, for each of the "th" tags I added a color and made it so they are all uppercase. And lastly, I added I tags to italicize all of the "th" tags.

BOOTSTRAP ADDED #2: Added in a social media link with my facebook, but the image is the actual link to my Facebook rather than just some words.